Change ManageOperators class to single region.
The ManageOperators class felt overloaded by working in multiple regions in a
single object instance. This changes class instances to be tied to a single
region provided on initialization. There was also some DRYing done for turning
the CyHy operators SSM value into a list since that is done in two different
methods. The cli code was updated to handle working with multiple regions to
compensate for reducing the functionality of the ManageOperators class.

I believe that this approach is a better method of implementing this
functionality. Exit codes may want another pass to be more informational than
everything was fine or something wasn't fine. As an example: if adding a user
you might have the SSH key part succeed while adding the user to the list of
active CyHy operators may fail. Although there is output from the class
alerting you, the exit code is the same if either of the two actions fails.